requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.13.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dill==0.3.6
Django==4.1.5
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
isort==5.11.4
kombu==5.2.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pylint==2.15.10
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.7
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
